Restructure viz function. Use recursive calls.
