+++ eegrasp/eegrasp.py
@@ -4,12 +4,11 @@ EEGRasP
 
 import numpy as np
 from pygsp2 import graphs, learning, graph_learning
-from tqdm import tqdm  # TODO: Does it belong here?
+from tqdm import tqdm
 from scipy import spatial
 import matplotlib.pyplot as plt
 import mne
-from mne.channels.layout import _auto_topomap_coords
-from .plotting import PlottingDefaults, _separate_kwargs
+from .viz import plot_graph
 
 
 class EEGrasp():
@@ -473,172 +472,10 @@ class EEGrasp():
 
             return W, Z
 
-    def plot_graph(self, graph=None, signal=None, coordinates=None, labels=None, montage=None,
-                   colorbar=True, axis=None, clabel='Edge Weights', kind='topoplot', show_names=True, **kwargs):
+    def plot(self, graph=None, signal=None, coordinates=None, labels=None, montage=None,
+             colorbar=True, axis=None, clabel='Edge Weights', kind='topoplot', show_names=True, **kwargs):
+        """ Plot graph over the eeg montage.
+        %(eegrasp.viz.plot_graph)s
         """
-        Plot the graph over the eeg montage.
-
-        Parameters
-        ----------
-        graph : PyGSP2 Graph object | None.
-            If `None` (default) the instance's graph will be used. If a `PyGSP2 Graph` object is passed, it will
-            be used to plot the graph.
-        signal : ndarray | list | None.
-            If `None` (default), vertices will have different size depending on their weighted degree and the edges
-            will have a different color depending on the connection strength between vertices. If a list or ndarray
-            is passed, the vertices will have a different color depending on the signal passed.
-        coordinates : ndarray | list | None.
-            If `None`, the instance's coordinates will be used.
-        labels : list | ndarray | None.
-            Labels to be plotted with vertices. If `None`, the instance's labels will be used.
-        montage : str | mne RawBase | mne EpochsBase | mne EvokedBase | None.
-            If `None`, the instance's coordenates will be used to build a custom montage. Since it
-            will only use the coordinateds tu build the custom montage, the sphere outline will
-            not be adjusted to contain the electrodes. If a string is used, it will try to build
-            a montage from the standard built-in mne library. If a ` mne DigiMontage` Class is
-            used it will plot the sensors using the given montage and set sphere parameter to
-            `None` when using `mne.viz.plot_sensors`.
-        colorbar : bool.
-            If True (default), a colorbar will be plotted.
-        clabel : str.
-            Label for the colorbar. Default is 'Edge Weights'.
-        axis : matplotlib axis object | None.
-            If `None` (default), a new figure will be created.
-        kind : str.
-            Kind of plot to use. Options are 'topoplot' and '3d'. Default is topoplot
-        %(pygsp2.plot)s
-        %(mne.viz.plot_sensors)s
-
-        Returns
-        -------
-        figure : matplotlib figure.
-            Figure object.
-        axis : matplotlib axis.
-            Axis object.
-
-        Notes
-        -----
-        Any argument from `mne.viz.plot_sensors` and `pygsp2.plot` can be passed to the function.
-        Default parameters can be found in `PlottingDefaults` class in `eegrasp/plotting.py`.
-
-        See Also
-        --------
-        * pygsp2 function `pygsp2.plotting.plot`
-        * mne function `mne.viz.plot_sensors`
-        """
-        # Load default values
-        default_values = PlottingDefaults()
-        original_kwargs = kwargs.copy()
-        # Add default values for plotting
-        kwargs = default_values.load_defaults(kwargs)
-        cmap = kwargs['cmap']
-
-        # Separate kwargs for pygsp2 and mne
-        pygsp_arg_list = self.graph.plot.__code__.co_varnames
-        mne_arg_list = mne.viz.plot_sensors.__code__.co_varnames
-
-        kwargs_pygsp_plot, kwargs = _separate_kwargs(kwargs, pygsp_arg_list)
-        kwargs_mne_plot, kwargs = _separate_kwargs(kwargs, mne_arg_list)
-
-        # Raise exemption if kwargs is not empty
-        if len(kwargs) > 0:
-            raise ValueError(f'Invalid arguments: {list(kwargs.keys())}')
-
-        # Handle variables if not passed
-        if graph is None:
-            graph = self.graph
-
-        if coordinates is None:
-            coordinates = self.coordinates
-
-        if axis is None:
-            fig = plt.figure()
-            if kind == 'topoplot':
-                axis = fig.add_subplot(111)
-            elif kind == '3d':
-                axis = fig.add_subplot(111, projection='3d')
-        else:
-            fig = axis.get_figure()
-
-        if labels is None:
-            labels = self.labels
-
-        if montage is None:
-            ch_pos = dict(zip(labels, coordinates))
-            montage = mne.channels.make_dig_montage(
-                ch_pos=ch_pos, coord_frame='head')
-        elif isinstance(montage, str):
-            try:
-                montage = mne.channels.make_standard_montage(montage)
-                labels = montage.ch_names
-                kwargs_mne_plot['sphere'] = None
-            except ValueError:
-                print(
-                    f'{montage} Montage not found. Creating custom montage based on self.coordenates...')
-                fig, axis = self.plot_graph(graph=graph, signal=signal, coordinates=coordinates, labels=labels, montage=montage,
-                                            colorbar=colorbar, axis=axis, clabel=clabel, kind=kind, show_names=show_names, **original_kwargs)
-                return fig, axis
-        else:
-            kwargs_mne_plot['sphere'] = None
-
-        if signal is None and 'edge_color' not in original_kwargs:
-
-            # Plot edge color depending on the edge weights
-            edge_weights = self.graph.get_edge_list()[2]
-
-            # if vertex size was not given, use weighted degree
-            if 'vertex_size' not in original_kwargs:
-                degree = np.array(graph.dw, dtype=float)
-                degree /= np.max(degree)
-                original_kwargs['vertex_size'] = degree
-
-            fig, axis = self.plot_graph(graph, edge_weights, coordinates, labels,
-                                        montage, colorbar, axis, clabel, kind, show_names, **original_kwargs)
-            return fig, axis
-
-        elif isinstance(signal, (list, np.ndarray)):
-
-            norm_signal = np.array(signal, dtype=float)
-            norm_signal -= np.min(norm_signal)
-            norm_signal /= np.max(norm_signal)
-            kwargs_pygsp_plot['edge_color'] = plt.cm.get_cmap(
-                cmap)(norm_signal)
-
-        if colorbar:
-            cbar = fig.colorbar(plt.cm.ScalarMappable(cmap=cmap),
-                                ax=axis, label=clabel)
-            cbar.set_ticks([0, 0.5, 1])
-            cbar.ax.set_yticklabels(
-                np.round([0, np.max(signal)/2, np.max(signal)], 2))
-
-        # Plot the montage
-        if kind == 'topoplot':
-
-            info = mne.create_info(labels, sfreq=250, ch_types="eeg")
-            info.set_montage(montage)
-
-            xy = _auto_topomap_coords(
-                info, None, True, to_sphere=True, sphere=kwargs_mne_plot['sphere'])
-            graph.set_coordinates(xy)
-            figure = mne.viz.plot_sensors(info, kind='topomap', show_names=show_names, ch_type='eeg',
-                                          axes=axis, show=False, **kwargs_mne_plot)
-            figure, axis = graph.plot(ax=axis, colorbar=colorbar,
-                                      **kwargs_pygsp_plot)
-
-        elif kind == '3d':
-
-            info = mne.create_info(labels, sfreq=250, ch_types="eeg")
-            info.set_montage(montage)
-            eeg_pos = montage.get_positions()['ch_pos']
-            eeg_pos = np.array([pos for _, pos in eeg_pos.items()])
-
-            dev_head_t = info["dev_head_t"]
-            eeg_pos = mne.transforms.apply_trans(dev_head_t, eeg_pos)
-            graph.set_coordinates(eeg_pos)
-
-            figure = mne.viz.plot_sensors(
-                info, kind='3d', show_names=True, axes=axis, show=False, **kwargs_mne_plot)
-            figure, axis = graph.plot(
-                ax=axis, colorbar=colorbar, **kwargs_pygsp_plot)
-
-        return (figure, axis)
+        return plot_graph(eegrasp=self, graph=graph, signal=signal, coordinates=coordinates, labels=labels, montage=montage,
+                          colorbar=colorbar, axis=axis, clabel=clabel, kind=kind, show_names=show_names, **kwargs)
