--- eegrasp/plotting.py
@@ -1,96 +1,0 @@
-"""Define default values and functions used for plotting function in eegrasp main module."""
-
-import dataclasses
-
-
-@dataclasses.dataclass
-class PlottingDefaults:
-    """Class containing default values and functions for eegrasp plotting functions.
-
-    Attributes
-    ----------
-    DEFAULT_CMAP : str.
-        Default colormap for plotting.
-    DEFAULT_VERTEX_COLOR : str.
-        Default color for the vertices.
-    DEFAULT_SPHERE : str.
-        Default sphere for plotting.
-    DEFAULT_ALPHAN : float.
-        Default alpha value for the vertices.
-    DEFAULT_VERTEX_SIZE : float.
-        Default size for the vertices.
-    DEFAULT_POINTSIZE : float.
-        Default point size for plotting.
-    DEFAULT_LINEWIDTH : float.
-        Default line width for plotting.
-    DEFAULT_EDGE_WIDTH : float.
-        Default edge width for plotting.
-    DEFAULT_EDGE_COLOR : str.
-        Default edge color for plotting.
-    DEFAULT_ALPHAV : float.
-        Default alpha value for the vertices.
-    """
-    DEFAULT_CMAP: str = 'Spectral_r'
-    DEFAULT_VERTEX_COLOR: str = 'teal'
-    DEFAULT_SPHERE: str = 'eeglab'
-    DEFAULT_ALPHAN: float = 0.5
-    DEFAULT_VERTEX_SIZE: float = 10.
-    DEFAULT_POINTSIZE: float = 0.5
-    DEFAULT_LINEWIDTH: float = 1.
-    DEFAULT_EDGE_WIDTH: float = 2.
-    DEFAULT_EDGE_COLOR: str = 'black'
-    DEFAULT_ALPHAV: float = 1.
-
-    def load_defaults(self, kwargs):
-        """Return dictionary with added default values for plotting functions if parameters have not
-        been set.
-
-        Parameters
-        ----------
-        kwargs : dict.
-            Dictionary containing the variables to be updated.
-
-        Returns
-        -------
-        kwargs : dict.
-            Dictionary with default values added.
-        """
-        for key, value in self.__dict__.items():
-            true_key = key.lower().rsplit('default_', maxsplit=1)[-1]
-            if true_key not in kwargs:
-                kwargs[true_key] = value
-        return kwargs
-
-
-def _update_locals(kwargs, local_vars):
-    """Update local variables with the kwargs dict.
-    Parameters
-    ----------
-    kwargs : dict.
-        Dictionary containing the variables to be updated.
-    local_vars : list.
-        List of local variables to be updated.
-    """
-    new_kwargs = kwargs.copy()
-    for key in kwargs.keys():
-        if key in local_vars:
-            local_vars[key] = kwargs[key]
-            del new_kwargs[key]
-    return new_kwargs
-
-
-def _separate_kwargs(kwargs, names):
-    """Separate kwargs into two dictionaries based on names on vars."""
-    var1 = {}
-    var2 = {}
-    for key in kwargs.keys():
-        if key in names:
-            var1[key] = kwargs[key]
-        else:
-            var2[key] = kwargs[key]
-    return var1, var2
-
-
-if __name__ == '__main__':
-    defaults = PlottingDefaults()
-    print(defaults.__dict__)
